separate queries into query.go
