+++ main.go
@@ -53,42 +53,11 @@ func main() {
 func getBlog(c *gin.Context) {
 	id := c.Param("id")
 	idInt, _ := strconv.Atoi(id)
-	myArticle, _ := queryBlog(idInt)
-	c.IndentedJSON(http.StatusOK, myArticle)
+	blogById, _ := queryBlog(idInt)
+	c.IndentedJSON(http.StatusOK, blogById)
 }
 
 func getAllBlogs(c *gin.Context) {
-	myArticle, _ := queryAllBlog()
-	c.IndentedJSON(http.StatusOK, myArticle)
-}
-
-func queryAllBlog() ([]Article, error) {
-	rows, err := db.Query("SELECT * FROM article WHERE published = true")
-	if err != nil {
-		return nil, err
-	}
-	var blogs []Article
-	for rows.Next() {
-		var blog Article
-		if err := rows.Scan(&blog.Id, &blog.Title, &blog.LastUpdate, &blog.Published,
-			&blog.ArticleType, &blog.Content, &blog.CreatedDate); err != nil {
-			return blogs, err
-		}
-		blogs = append(blogs, blog)
-	}
-	if err != nil {
-		return nil, err
-	}
-	return blogs, nil
-}
-
-func queryBlog(id int) (Article, error) {
-	blog := Article{}
-	err := db.QueryRow("SELECT * FROM article WHERE id = $1", id).Scan(
-		&blog.Id, &blog.Title, &blog.LastUpdate, &blog.Published,
-		&blog.ArticleType, &blog.Content, &blog.CreatedDate)
-	if err != nil {
-		return blog, err
-	}
-	return blog, nil
+	allBlogs, _ := queryAllBlog()
+	c.IndentedJSON(http.StatusOK, allBlogs)
 }
